refactor(agent-service): rename types/wire.ts to types/dto.ts
Pure file rename plus import-path updates; no type or behavior changes.
